Microservices: add timestamp to go-service logs
